+++ CrazyRandomSword.h
@@ -24,7 +24,7 @@ public:
         */
 
         srand(time(0));                                     // seed random number generator
-        hitPoints = static_cast<int>(rand() % 94 + 7);       // set hitPoints to random number between 7 - 100
+        hitPoints = rand() % 94 + 7;       // set hitPoints to random number between 7 - 100
     }
 
     virtual ~CrazyRandomSword() {};
